Give TemplateDescriptions proper equals() and hashCode() (#312)
* Give TemplateDescriptions proper equals() and hashCode()

These types are frequently used in Sets. Accordingly, they'll be hashed, and so should have proper hashCode() and equals() implementations.

The implementations I chose were a best-effort guess as to the identity of the models. ClassInfo was particularly challenging as it has a mutable field and I found it challenging to figure out the difference between the various types of names.

* fixup! Give TemplateDescriptions proper equals() and hashCode()

* fixup! Give TemplateDescriptions proper equals() and hashCode()